refactor: separate files
Separate the single plugin into a few files, matching the names of the tests for them.
