+++ test/plugin.spec.js
